Move integer promotion traits to use with math functions
